+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.4.0] -2020-03-04
+### Fixed
+- New linter errors.
+### Changed
+- Use ErrStack instead Error in state.
+- Error's workflow of state.
 
 ## [0.3.4] - 2019-09-17
 ### Fixed